## **Description**

THIS PR IMPROVED THE UNLOCK TIME (AFTER USER PRESSES UNLOCK UNTIL THE WALLET VIEW), OF 18 SECONDS (debug mode)
Previous unlock duration: ~25 seconds
New unlock duration: ~7 seconds
And from ~10 seconds to ~4 seconds 
(Android QA build)! 
Previous unlock duration: ~10 seconds
New unlock duration: ~4 seconds

We will be able to drop our baseline on our performance E2E testing, seeing results in this [build](https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/b59e9827-b897-4769-8fee-9b0c50a65992?tab=workflows) time that it takes for a cold app start to the wallet view, reducing 30% (1.4 seconds) of the time, compared to last main build  (Android)

This PR patch scure-bip39, eth-hd-keyring and keyring controller to use the native aes module instead.

This patch will be replaced for ongoing work on Keyring Controller,  key-tree, and eth-hd-keyring

These are the next steps that need to be taken to ship this:
* Reviews & merge https://github.com/MetaMask/accounts/pull/100
* Release [key-tree](https://github.com/MetaMask/key-tree) last version
* Update https://github.com/MetaMask/accounts/pull/102 with key-tree last version
* Refine Keyring Controller with the necessary changes to use the new eth-hd-keyring or create a keyring builder from mobile when [concat](https://github.com/MetaMask/core/blob/5f6b020693b6497c40728e80c281b5cc3028c542/packages/keyring-controller/src/KeyringController.ts#L649) issue is fixed to be able to override hd-keyring from the Keyring Controller consumer
* We can use this bump [PR](https://github.com/MetaMask/metamask-mobile/pull/12339) of KeyringController to bring those changes to mbile



PR smoke e2e pipeline: https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/2fe81597-0389-43e3-87f1-b74b905707f2?tab=workflows

## **Related issues**

Fixes:

## **Manual testing steps**

1. Go to this page...
2.
3.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the before and after of your change. -->

### **Before**

(DEBUG)
https://github.com/user-attachments/assets/0d58ae22-84bb-463d-8948-da06649cd303

(QA)

https://github.com/user-attachments/assets/706ae004-6c12-44a6-ad3f-b976fe0d7333



### **After**

(DEBUG)
https://github.com/user-attachments/assets/009f2ef4-5239-4944-979a-bd57adcc1d26

(QA)
https://github.com/user-attachments/assets/3e264ce5-0842-480b-b9bf-cfd8c6d34258



## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Contributor Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile Coding Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)). Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described in the ticket it closes and includes the necessary testing evidence such as recordings and or screenshots.